+++ system/sysctl/sysctl_darwin.go
@@ -1,4 +1,4 @@
-package ek
+package sysctl
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,5 +7,15 @@ package ek
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// VERSION is current ek package version
-const VERSION = "13.4.0"
+import (
+	"os/exec"
+	"strings"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// getKernelParam reads kernel parameter from procfs
+func getKernelParam(param string) (string, error) {
+	output, err := exec.Command(binary, "-n", param).Output()
+	return strings.Trim(string(output), "\n\r"), err
+}