+++ system/sysctl/sysctl_linux.go
@@ -1,4 +1,4 @@
-package ek
+package sysctl
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,5 +7,25 @@ package ek
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// VERSION is current ek package version
-const VERSION = "13.4.0"
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/essentialkaos/ek/v13/path"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// getKernelParam reads kernel parameter from procfs
+func getKernelParam(param string) (string, error) {
+	p, err := os.ReadFile(path.Clean(path.Join(
+		procFS, strings.ReplaceAll(param, ".", "/"),
+	)))
+
+	if err != nil {
+		return "", fmt.Errorf("Can't read parameter %q: %w", param, err)
+	}
+
+	return strings.Trim(string(p), "\n\r"), nil
+}