+++ system/sysctl/sysctl_stubs.go
@@ -1,4 +1,8 @@
-package ek
+//go:build !linux && !darwin
+// +build !linux,!darwin
+
+// Package sysctl provides methods for reading kernel parameters
+package sysctl
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,5 +11,17 @@ package ek
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// VERSION is current ek package version
-const VERSION = "13.4.0"
+// Get returns kernel parameter value as a string
+func Get(param string) (string, error) {
+	return "", nil
+}
+
+// GetI returns kernel parameter value as an int
+func GetI(param string) (int, error) {
+	return 0, nil
+}
+
+// GetI64 returns kernel parameter value as an int64
+func GetI64(param string) (int64, error) {
+	return 0, nil
+}